refactor: adjust authorization plugin based on mentor review cy-9
