+++ apps/backend/src/modules/users/libs/types/types.ts
@@ -1,4 +1,6 @@
 export {
+	type UserGetAllItemResponseDto,
+	type UserGetAllResponseDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
 	type UserSignUpRequestDto,
