+++ apps/backend/src/plugins/authorization/libs/types/types.ts
@@ -1,0 +1,5 @@
+export {
+	AuthorizationError,
+	ErrorMessage,
+	type UserSignUpResponseDto,
+} from "shared";
